refactor: ミニ結果のXP演出をフックに分離
